HBX-1598: Rename class 'org.hibernate.tool.hbm2x.ArtifactCollector' to 'org.hibernate.tool.internal.exporter.DefaultArtifactCollector'
